Log::wtf() can now force a terminate.
By default you only get the sudden abort when assertions are enabled.
This setting allows forcing the termination even when assertions
are disabled with NDEBUG.

The fact that you should only call wtf() if your program is in such
a terrible failure state that it should suddenly abort however remains.