Put gcc/clang back to normal export.
That is, export everything. Because the only way I've ever found to make the
customsink error go away on gcc/clang is to export the entire class, which msvc
doesn't like because of C++ runtime exports. So let's just go with the old way,
which is assume everything is exported unless you're building for msvc.